Refactor and enhance world rendering components
- Removed unused logger imports in usePointerLockInput.ts
- Added a blank line for better readability in BuildingDrawer.tsx
- Imported chunkKey from chunkHelpers in useMeshedChunks.ts
- Adjusted variable names for clarity in drones.ts
- Simplified imports in tickDrones.ts
- Imported chunkKey in meshingScheduler.ts and removed redundant function definition
- Reordered imports in createSimBridge.ts for consistency
- Updated default UI state in store.ts to include haulerCount and outposts
- Modified player physics tests to use new ground height calculation
- Updated protocol tests to include haulerCount and outposts
- Enhanced error handling tests for sim-bridge to include new state properties
- Added new FrontierFillRenderer and VoxelLayerInstanced components for improved voxel rendering
- Introduced VoxelLayerMeshed component to manage meshed voxel rendering with LOD support
