+++ src/components/UI.tsx
@@ -3,8 +3,8 @@ import React, { useState } from "react";
 import type { ViewMode } from "../types";
 import { useUiStore } from "../ui/store";
 import { SettingsModal } from "./SettingsModal";
-import { Hud } from "./ui/Hud";
 import { BuildingDrawer } from "./ui/BuildingDrawer";
+import { Hud } from "./ui/Hud";
 import { Reticle } from "./ui/Reticle";
 import { ShopModal } from "./ui/ShopModal";
 import { TitlePanel } from "./ui/TitlePanel";
