+++ tests/protocol.test.ts
@@ -50,12 +50,14 @@ describe("protocol payloads", () => {
         credits: 0,
         prestigeLevel: 1,
         droneCount: 0,
+        haulerCount: 0,
         miningSpeedLevel: 1,
         moveSpeedLevel: 1,
         laserPowerLevel: 1,
         minedBlocks: 0,
         totalBlocks: 0,
         upgrades: {},
+        outposts: [],
         nextCosts: { drone: 100 },
       },
       stats: { simMs: 0.1, backlog: 0 },
