+++ tests/sim-bridge-error-handling.test.ts
@@ -144,12 +144,14 @@ describe("SimBridge worker error handling with retry", () => {
         credits: 0,
         prestigeLevel: 1,
         droneCount: 0,
+        haulerCount: 0,
         miningSpeedLevel: 1,
         moveSpeedLevel: 1,
         laserPowerLevel: 1,
         minedBlocks: 0,
         totalBlocks: 0,
         upgrades: {},
+        outposts: [],
       },
       stats: { simMs: 0, backlog: 0 },
     });
