+++ tests/sim-bridge.test.ts
@@ -52,12 +52,14 @@ describe("sim bridge gating", () => {
         credits: 0,
         prestigeLevel: 1,
         droneCount: 0,
+        haulerCount: 0,
         miningSpeedLevel: 1,
         moveSpeedLevel: 1,
         laserPowerLevel: 1,
         minedBlocks: 0,
         totalBlocks: 0,
         upgrades: {},
+        outposts: [],
       },
       stats: { simMs: 0, backlog: 0 },
     });
